Rename linked_list_program to Even_Odd_separator (using linked list)/linked_list_program.cpp
This Program will ask the user to enter a couple of numbers, after that it will convert those numbers into a list. it will separate the even numbers from the entered value as well as odd numbers from the entered value and will print them on the screen. After that, the program will also merge those values, printing the odd numbers first and later the even numbers. 

The Node class will create the Nodes (elements of the list).